refactor: rename controller mappers for clarity
